requirements.txt
flask
flask-socketio
websockets
asyncio
kafka-python
gunicorn
websocket-client
python-socketio
requests
gevent
pytz
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
flask-cors
Flask-Migrate==4.0.4
cloudinary==1.36.0
eventlet>=0.33
gevent-websocket
eventlet
flask-jwt-extended
python-dateutil
pytz